scripts/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
openai
pyyaml
